requirements_full.txt
# ============================================================
# image_3d_transfiguration v3 - FULL VERSION Requirements
# Python 3.9 기준 / GPU(CUDA 11.8) Deep3D Stack 안정 조합
# ============================================================

# ---------------- Core ----------------
numpy>=1.23
opencv-python>=4.8.0
pyyaml>=6.0
tqdm>=4.66.0
scikit-image>=0.22.0

# ---------------- 3D ------------------
open3d>=0.18.0
trimesh>=4.4.0
pygltflib>=1.16.0

# ---------------- Torch ----------------
# pip install → indices 없이 설치하면 CPU-only가 설치되므로
# 반드시 아래 명령으로 CUDA 11.8 버전 설치:
#   pip install torch==2.1.0+cu118 torchvision==0.16.0+cu118 torchaudio==2.1.0+cu118 \
#       --extra-index-url https://download.pytorch.org/whl/cu118

torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

# ---------------- pytorch3d ----------------
# pip 명령 (Python 3.9 / Torch 2.1.0 / CUDA 11.8)
#   pip install pytorch3d==0.7.4 \
#       --extra-index-url https://dl.fbaipublicfiles.com/pytorch3d/packaging/wheels/cu118

pytorch3d==0.7.4

# ---------------- Depth (HF) ----------------
transformers>=4.38.0
accelerate>=0.30.0
huggingface-hub>=0.23.0
Pillow>=10.0.0

# ---------------- Zoe/DepthAnything ----------------
# ZoeDepth: 직접 clone 필요 (pip로 안됨)
#   git clone https://github.com/isl-org/ZoeDepth.git
#   pip install -e ZoeDepth
#
# DepthAnythingV2는 HF모델 기반으로 동작

# ---------------- SAM / YOLO ----------------
# Meta SAM은 git 설치 필요
#   pip install git+https://github.com/facebookresearch/segment-anything.git
#
ultralytics>=8.0.200

# ---------------- Utils ---------------------
matplotlib>=3.8.0
loguru>=0.7.0
streamlit>=1.32.0